+++ NEWS.md
@@ -3,6 +3,13 @@ title: "NEWS"
 output: html_document
 ---
 
+## incidentally 1.0.2
+
+* added reference to Neal (2022) for incidence.from.congress()
+* added reference to Neal (2023) for incidence.from.adjacency()
+* `incidence.from.congress()` includes bill attributes indicating the number of (co-)sponsors from each party
+* fixed minor bugs introduced by `igraph 1.4.0`
+
 ## incidentally 1.0.1
 
 * fixed bug in `incidence.from.congress()` when bills do not have a sponsor